LayoutTests/fast/forms/datetime/datetime-input-type.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('Tests for writing and reading .type property of HTMLInputElement.');

var input = document.createElement('input');
document.body.appendChild(input);

function check(value, expected)
{
    input.type = value;
    if (input.type == expected)
        testPassed('input.type for "' + value + '" is correctly "' + input.type + '".');
    else
        testFailed('input.type for "' + value + '" is incorrectly "' + input.type + '", should be "' + expected + '".');
}

check("datetime", "datetime");
check("DATETIME", "datetime");
check("DateTime", "datetime");
check("dateTime", "datetime");
check(" datetime ", "text");
check("date-time", "text");
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
