LayoutTests/fast/forms/datetime/datetime-input-type-expected.txt

Tests for writing and reading .type property of HTMLInputElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.type for "datetime" is correctly "datetime".
PASS input.type for "DATETIME" is correctly "datetime".
PASS input.type for "DateTime" is correctly "datetime".
PASS input.type for "dateTime" is correctly "datetime".
PASS input.type for " datetime " is correctly "text".
PASS input.type for "date-time" is correctly "text".
PASS successfullyParsed is true

TEST COMPLETE
